pyproject.toml
[tool.poetry]
authors = ["Darkflame72 <leon@bowie-co.nz>"]
description = "A Minecraft discord bot"
license = "AGPL-3.0-or-later"
name = "obsidion"
version = "1.0.0"

[tool.poetry.dependencies]
Babel = "^2.9.0"
aiodns = "^2.0.0"
aiohttp = "^3.6.2"
aiohypixel = "^0.2.1"
aioredis = "^1.3.1"
alembic = "^1.4.3"
asyncpg = "^0.21.0"
asyncrcon = "^1.1.4"
"discord.py" = "^1.5.0"
feedparser = "^6.0.1"
fuzzywuzzy = {extras = ["speedup"], version = "^0.18.0"}
lxml = "^4.5.2"
mcsrvstats = "^0.1.1"
pydantic = {extras = ["dotenv"], version = "^1.7.3"}
python = "^3.8"
uvloop = "^0.14.0"
SQLAlchemy = "1.4.0b1"
asyncpixel = "^1.0.1"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
Pygments = "^2.7.2"
black = "^20.8b1"
coverage = {extras = ["toml"], version = "^5.3"}
flake8 = "^3.8.4"
mypy = "^0.790"
nox = "^2020.8.22"
nox-poetry = "^0.6.0"
pre-commit = "^2.8.2"
pre-commit-hooks = "^3.3.0"
pytest = "^6.1.2"
pytest-asyncio = "^0.14.0"
reorder-python-imports = "^2.3.6"
safety = "^1.9.0"
typeguard = "^2.9.1"
flake8-bugbear = "^20.11.1"
pep8-naming = "^0.11.1"
flake8-builtins = "^1.5.3"
flake8-bandit = "^2.1.2"
flake8-eradicate = "^1.0.0"
flake8-comprehensions = "^3.3.1"
flake8-import-order = "^0.18.1"

[tool.coverage.paths]
source = ["bot"]

[tool.coverage.run]
branch = true
source = ["bot"]

[tool.coverage.report]
fail_under = 100
show_missing = true

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.1"]
